+++ libs/extract/src/index.ts
@@ -1,4 +1,5 @@
 export * from './lib/alert'
+export * from './lib/badge'
 export * from './lib/button'
 export * from './lib/datepicker'
 export * from './lib/dropdown'