+++ libs/extract/src/lib/dropdown/index.ts
@@ -8,6 +8,7 @@ export {
   DropdownListener,
   ElementProps,
   DropdownArgs,
+  DropdownTexts,
 } from './types'
 
 export { dropdownValues, optionValues } from './defaultValues'